# app/api/v1/wallets/withdrawal/crypto/route.ts
import {
  checkIdempotency,
  storeIdempotencyResponse,
} from "@/lib/middlewares/idempotency";
import { unauthorizedResponse, validateJWT } from "@/lib/middlewares/jwt-auth.";
import { checkRateLimit, rateLimiters } from "@/lib/middlewares/rate-limit";
import { type NextRequest, NextResponse } from "next/server";

import { logAuditEvent } from "@/lib/services/audit.service";
import {
  initiateCryptoWithdrawal,
  initiateFiatDeposit,
  verifySufficientFunds,
} from "@/lib/services/wallets.service";
import { supabaseAdmin } from "@/lib/supabase/admin";
import { createSupabaseClient } from "@/lib/supabase/client";
import { withdrawCryptoSchema } from "@/lib/validators/wallet.schema";

export async function POST(req: NextRequest) {
  let user: any = null;

  try {
    // Step 1: Authenticate
    const authResult = await validateJWT(req);
    user = authResult.user;
    if (!user)
      return unauthorizedResponse(
        authResult.error || "Authentication required"
      );

    // Step 2: Rate limit
    const rateLimitCheck = await checkRateLimit(
      req,
      rateLimiters.withdrawal,
      user.id
    );
    if (rateLimitCheck.limited) return rateLimitCheck.response!;

    const body = await req.json();
    const idempotencyKey = req.headers.get("x-request-id");

    // Step 3: Idempotency
    if (idempotencyKey) {
      const idempotencyResult = await checkIdempotency(
        req,
        user.id,
        "/api/v1/wallets/withdrawal/crypto",
        body
      );

      if (idempotencyResult.isDuplicate && idempotencyResult.existingResponse) {
        return NextResponse.json(idempotencyResult.existingResponse, {
          status: 200,
          headers: { "X-Idempotency-Status": "HIT" },
        });
      }
    }

    // Step 4: Validate input
    const validation = withdrawCryptoSchema.safeParse(body);
    if (!validation.success) {
      return NextResponse.json(
        {
          error: "Validation failed",
          code: "VALIDATION_ERROR",
          details: validation.error.flatten().fieldErrors,
        },
        { status: 400 }
      );
    }

    const {
      asset_id,
      amount: amountRaw,
      to_address,
      network,
      two_factor_code,
    } = validation.data;
    const amount =
      typeof amountRaw === "string" ? Number(amountRaw) : amountRaw;
    if (!Number.isFinite(amount) || amount <= 0) {
      return NextResponse.json(
        { error: "Invalid amount", code: "VALIDATION_ERROR" },
        { status: 400 }
      );
    }

    // Step 5: Verify 2FA
    const authHeader = req.headers.get("authorization");
    const token = authHeader?.replace("Bearer ", "");
    if (!token) return unauthorizedResponse("Authentication token required");

    const supabaseClient = createSupabaseClient(token);
    const { data: userData } = await supabaseClient
      .from("users")
      .select("two_factor_enabled")
      .eq("id", user.id)
      .single();

    if (userData?.two_factor_enabled) {
      if (!two_factor_code || two_factor_code.length !== 6) {
        return NextResponse.json(
          { error: "2FA verification required", code: "2FA_REQUIRED" },
          { status: 403 }
        );
      }
    }

    // Step 6: Get asset info
    const { data: asset } = await supabaseAdmin
      .from("crypto_assets")
      .select("*")
      .eq("id", asset_id)
      .single();

    if (!asset)
      return NextResponse.json(
        { error: "Asset not found", code: "ASSET_NOT_FOUND" },
        { status: 404 }
      );
    if (amount < Number(asset.min_withdrawal)) {
      return NextResponse.json(
        {
          error: `Minimum withdrawal is ${asset.min_withdrawal} ${asset.ticker}`,
          code: "MIN_WITHDRAWAL",
        },
        { status: 400 }
      );
    }

    // Step 7: Get wallet info
    const wallet = await initiateFiatDeposit({
      userId: user.id,
      amount: "0",
      currency: asset.ticker,
      paymentMethod: "internal",
    });

    // Step 8: Verify sufficient funds
    const { sufficient, availableBalance } = await verifySufficientFunds(
      user.id,
      wallet.id,
      amount.toString(),
      "crypto"
    );
    if (!sufficient) {
      return NextResponse.json(
        {
          error: "Insufficient balance",
          code: "INSUFFICIENT_FUNDS",
          availableBalance,
        },
        { status: 400 }
      );
    }

    // Step 9: Initiate withdrawal (reserves funds internally)
    const withdrawalResult = await initiateCryptoWithdrawal({
      userId: user.id,
      requestId: idempotencyKey || `req-${Date.now()}`,
      assetTicker: asset.ticker,
      amount: amount.toString(),
      destinationAddress: to_address,
      network,
      twoFaCode: two_factor_code || "",
    });

    // Step 10: Record transaction
    const { data: transaction, error: txError } = await supabaseAdmin
      .from("crypto_transactions")
      .insert({
        user_id: user.id,
        crypto_wallet_id: wallet.id,
        asset_id,
        transaction_type: "withdrawal",
        amount: amount.toString(),
        fee: Number(asset.withdrawal_fee),
        net_amount: amount.toString(),
        to_address,
        network,
        status: "processing",
        two_factor_verified: !!two_factor_code,
        idempotency_key: idempotencyKey,
      })
      .select()
      .single();

    if (txError) throw txError;

    await supabaseAdmin
      .from("crypto_transactions")
      .update({
        status: "completed",
        tx_hash: `0x${Math.random().toString(16).substring(2, 66)}`,
        completed_at: new Date().toISOString(),
      })
      .eq("id", transaction.id);

    // Step 11: Log audit
    await logAuditEvent({
      user_id: user.id,
      event_type: "crypto_withdrawal_initiated",
      event_category: "wallet",
      severity: "info",
      event_data: { amount, asset: asset.ticker, to_address, network },
      metadata: { transaction_id: transaction.id },
    });

    const response = {
      message: "Withdrawal initiated successfully",
      transaction: {
        id: transaction.id,
        amount: amount.toString(),
        fee: Number(asset.withdrawal_fee),
        asset: asset.ticker,
        to_address,
        status: "processing",
      },
    };

    if (idempotencyKey)
      await storeIdempotencyResponse(idempotencyKey, 201, response);

    return NextResponse.json(response, {
      status: 201,
      headers: { "X-Idempotency-Status": idempotencyKey ? "STORED" : "N/A" },
    });
  } catch (error) {
    console.error("[v0] Crypto withdrawal error:", error);
    return NextResponse.json(
      { error: "Internal server error", code: "INTERNAL_ERROR" },
      { status: 500 }
    );
  }
}