# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FireTrade Mobile Application API
  version: 1.0.0
  description: |
    MVP Backend API for FireTrade - A Crypto Exchange and Bill Payment Service
    supporting NGN fiat pairs and digital bill payments in Nigeria.
  contact:
    name: FireTrade API Support
    email: api@firetrade.ng
servers:
  - url: http://localhost:3002/api/v1/
    description: Production server
  - url: https://staging-api.firetrade.ng/api/v1
    description: Staging server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token obtained from authentication endpoints
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid request parameters"
        details:
          type: object
          additionalProperties: true
...
    WithdrawCryptoRequest:
      type: object
      required:
        - asset_ticker
        - amount
        - destination_address
        - network
      properties:
        asset_ticker:
          type: string
          example: "BTC"
        amount:
          type: string
          pattern: '^\d+\.\d{1,10}$'
          example: "0.05"
        destination_address:
          type: string
          example: "bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh"
        network:
          type: string
          example: "bitcoin"
        two_fa_code:
          type: string
          pattern: "^[0-9]{6}$"
          example: "123456"
    WithdrawCryptoResponse:
      type: object
      properties:
        withdrawal_id:
          type: string
          format: uuid
          example: "850e8400-e29b-41d4-a716-446655440005"
        status:
          type: string
          enum: [pending_approval, processing, completed, failed, rejected]
          example: "pending_approval"
        estimated_fee:
          type: string
          example: "0.0005"
        estimated_completion:
          type: string
          format: date-time
...
    TradeRequest:
      type: object
      required:
        - pair
        - type
        - amount
      properties:
        pair:
          type: string
          pattern: "^[A-Z]{3,4}-[A-Z]{3}$"
          example: "BTC-NGN"
          description: "Trading pair in format BASE-QUOTE (e.g., BTC-NGN)"
        type:
          type: string
          enum: [BUY, SELL]
          example: "BUY"
          description: "BUY or SELL order type"
        amount:
          type: string
          pattern: '^\d+\.\d{2,10}$'
          example: "1000000.00"
          description: "Amount in quote currency (NGN for BTC-NGN pair)"
    TradeResponse:
      type: object
      properties:
        trade_id:
          type: string
          format: uuid
          example: "950e8400-e29b-41d4-a716-446655440006"
        pair:
          type: string
          example: "BTC-NGN"
        type:
          type: string
          example: "BUY"
        amount_base:
          type: string
          example: "0.0117647059"
          description: "Amount of base currency (BTC)"
        amount_quote:
          type: string
          example: "1000000.00"
          description: "Amount of quote currency (NGN)"
        execution_price:
          type: string
          example: "85000000.00"
        fee:
          type: string
          example: "2500.00"
        status:
          type: string
          enum: [pending, executed, failed, cancelled]
          example: "executed"
        executed_at:
          type: string
          format: date-time
...
    BillPaymentRequest:
      type: object
      required:
        - category_id
        - amount
        - fiat_wallet_id
        - recipient_details
      properties:
        category_id:
          type: string
          format: uuid
          example: "a50e8400-e29b-41d4-a716-446655440007"
        amount:
          type: string
          pattern: '^\d+\.\d{2}$'
          example: "5000.00"
        fiat_wallet_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        recipient_details:
          type: object
          description: "Provider-specific JSON object"
          example:
            provider: "MTN"
            phone: "08012345678"
          additionalProperties: true
    BillPaymentResponse:
      type: object
      properties:
        payment_id:
          type: string
          format: uuid
          example: "b50e8400-e29b-41d4-a716-446655440008"
        category:
          type: string
          example: "Airtime"
        amount_ngn:
          type: string
          example: "5000.00"
        status:
          type: string
          enum: [pending, processing, completed, failed]
          example: "completed"
        provider_reference_id:
          type: string
          example: "MTN-REF-20250512-789456"
        created_at:
          type: string
          format: date-time
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: true
      description: Unique request identifier for idempotency (UUID v4)
      schema:
        type: string
        format: uuid
        example: "c50e8400-e29b-41d4-a716-446655440009"
...
paths:
...
  /wallets/withdrawal/crypto:
    post:
      tags:
        - Wallet Management
      summary: Initiate crypto withdrawal
      description: Initiates crypto withdrawal request (requires 2FA confirmation)
      parameters:
        - $ref: "#/components/parameters/XRequestID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawCryptoRequest"
      responses:
        "201":
          description: Withdrawal initiated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawCryptoResponse"
        "400":
          description: Invalid request data or insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized or invalid 2FA code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Duplicate request (idempotency check failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /exchange/trade:
    post:
      tags:
        - Crypto Exchange
      summary: Place a spot market order
      description: Executes a spot market order (BUY or SELL) with idempotency support
      parameters:
        - $ref: "#/components/parameters/XRequestID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeRequest"
      responses:
        "201":
          description: Trade executed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResponse"
        "400":
          description: Invalid request data or insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Duplicate request (idempotency check failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Trading temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /bills/pay:
    post:
      tags:
        - Bill Payments
      summary: Execute bill payment
      description: Executes a bill payment transaction with idempotency support
      parameters:
        - $ref: "#/components/parameters/XRequestID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillPaymentRequest"
      responses:
        "201":
          description: Bill payment successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillPaymentResponse"
        "400":
          description: Invalid request data or insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Duplicate request (idempotency check failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Bill payment service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"